All Dataset downloads should be logged
Resolves #115

The solution was to add a new table and track each time a user downloads a dataset via /api/v1/datasets/:id/archive.  This is the only way a user may download a dataset from the archive service.  There is a new Django Admin interface for downloading Dataset Download activity.  This allows the Django admins to filter to download logs by date range or user and to select all the available records to a csv.