Able to submit dataset without uploading a file
Issue #176

This involved expanding permissions and adding the 'archive' field to the check for permissions.  Now if a user tries to submit a dataset without uploading a file 'archive' will show up as a missingRequiredfield. This will also need some front end work to highlight upload area as missing or something like that.